Shapefile Uploads/requirements.txt
geopandas>=0.14.0
supabase>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
shapely>=2.0.0
pyproj>=3.6.0